requirements.txt
asgiref
astroid
autopep8
boto3
botocore
certifi
chardet
Click==7.0
coreapi
coreschema
coverage
Deprecated
dj-database-url
Django==2.1.7
django-autoslug
django-cors-headers
django-environ
django-extensions
-e git+https://github.com/wisnuprama/django-jet.git@428ea5b46dbc86597b48578683e499797b101aee#egg=django_jet
django-rq
django-storages
djangorestframework
djangorestframework-camel-case
djangorestframework-csv
djangorestframework-simplejwt
docutils
elastic-transport
elasticsearch
factory-boy
Faker
fuzzywuzzy
gunicorn
idna
isort
itypes
Jinja2
jmespath
lazy-object-proxy
Markdown
MarkupSafe
mccabe
Pillow
psycopg2-binary
pycodestyle
Pygments
PyJWT
pylint
python-dateutil
python-http-client
pytz
redis
requests
rq==1.2.0
s3transfer
sendgrid
sentry-sdk
setuptools
six
sqlparse
text-unidecode
typed-ast
unittest-xml-reporting
typing-extensions
unicodecsv
uritemplate
urllib3
wheel
whitenoise
wrapt
celery
importlib-metadata==4.8.3